requirements.txt
flask
flask-smorest
Flask-SQLAlchemy
Flask-Migrate
passlib
requests
python-dotenv
marshmallow
psycopg2
psycopg2-binary